We're moving away from the hacky way in which we were trying to figure out which users we already know about which we didn't via userhash
We're also moving away from trying to do updates on users we already
know about in the user table.  Instead, we now have a user_created
interface for just the part of this that involves keeping track of the
creation time of a user.

And now that we more reliably have that, we've added csv_reader.py to
upload the Slack's member table, and we're reporting on account creation
date in the user report